refactor: Lotto 클래스 생성자 중복 검사 로직 추가
Lotto 클래스 생성자 중복 검사 로직 추가 및 패키지 수정
