--- src/main/java/lotto/controller/util/ErrorMessage.java
@@ -1,7 +1,0 @@
-package lotto.controller.util;
-
-public class ErrorMessage {
-    public static final String INPUT_NOT_NUMBER = "[ERROR] 로또 번호는 숫자만 가능합니다.";
-    public static final String NUMBER_DUPLICATED = "[ERROR] 로또 번호는 중복될 수 없습니다.";
-    public static final String WRONG_LOTTO_NUMBER_LENGTH = "[ERROR] 로또 번호는 중복될 수 없습니다.";
-}
