+++ src/main/java/lotto/model/Lotto.java
@@ -1,13 +1,16 @@
 package lotto.model;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import lotto.configuration.ErrorMessage;
 
 public class Lotto {
     private final List<Integer> numbers;
 
     public Lotto(List<Integer> numbers) {
         validate(numbers);
+        isDuplicated(numbers);
         this.numbers = numbers;
     }
 
@@ -18,6 +21,13 @@ public class Lotto {
     }
 
     // TODO: 추가 기능 구현
+    private void isDuplicated(List<Integer> numbers) {
+        HashSet<Integer> nonDuplicateNumbers = new HashSet<>(numbers);
+        if (nonDuplicateNumbers.size() != numbers.size()) {
+            throw new IllegalArgumentException(ErrorMessage.NUMBER_DUPLICATED);
+        }
+    }
+
     public List<Integer> getNumbers() {
         return new ArrayList<>(numbers);
     }
